refactor: 메인페이지 canonical 처리
- 따로 컴포넌트로 분리화함 메타데이터와 use client때문
